packages/backend/requirements.txt
mcp
langchain-mcp-adapters
langgraph
langchain[openai]
fastapi
uvicorn[standard]
aiohttp
pydantic
python-dotenv